ResourceLoader and some Entity2D improvements

+++ src/main/java/com/shc/silenceengine/scene/Scene.java
@@ -1,8 +1,6 @@
 package com.shc.silenceengine.scene;
 
 import com.shc.silenceengine.graphics.Batcher;
-import com.shc.silenceengine.graphics.opengl.Program;
-import com.shc.silenceengine.graphics.opengl.Texture;
 
 /**
  * @author Sri Harsha Chilakapati
@@ -11,27 +9,22 @@ public final class Scene extends SceneNode
 {
     public Scene()
     {
-//        Program.CURRENT = null;
-//        Texture.CURRENT = null;
-//
-//        Program.DEFAULT.use();
-//        Texture.EMPTY.bind();
     }
 
-    public void preUpdate(double delta)
+    public void preUpdate(float delta)
     {
     }
 
-    public void preRender(double delta, Batcher batcher)
+    public void preRender(float delta, Batcher batcher)
     {
     }
 
-    public void update(double delta)
+    public void update(float delta)
     {
         updateChildren(delta);
     }
 
-    public void render(double delta, Batcher batcher)
+    public void render(float delta, Batcher batcher)
     {
         renderChildren(delta, batcher);
     }
